move events fixture to shared karma folder
